refactor: replace tag modal with dropdown menus
