+++ src/screens/Ingredients/AllIngredientsScreen.js
@@ -26,7 +26,7 @@ import HeaderWithSearch from "../../components/HeaderWithSearch";
 import { useTabMemory } from "../../context/TabMemoryContext";
 import { MaterialIcons } from "@expo/vector-icons";
 import { useTheme } from "react-native-paper";
-import TagFilterModal from "../../components/TagFilterModal";
+import TagFilterMenu from "../../components/TagFilterMenu";
 import { BUILTIN_INGREDIENT_TAGS } from "../../constants/ingredientTags";
 import { getAllTags } from "../../storage/ingredientTagsStorage";
 
@@ -200,7 +200,6 @@ export default function AllIngredientsScreen() {
   const [searchDebounced, setSearchDebounced] = useState("");
   const [navigatingId, setNavigatingId] = useState(null);
   const [selectedTagIds, setSelectedTagIds] = useState([]);
-  const [tagModalVisible, setTagModalVisible] = useState(false);
   const [availableTags, setAvailableTags] = useState([]);
 
   const didSetTabRef = useRef(false);
@@ -391,7 +390,13 @@ export default function AllIngredientsScreen() {
         searchValue={search}
         setSearchValue={setSearch}
         onMenu={() => navigation.navigate("GeneralMenu")}
-        onFilter={() => setTagModalVisible(true)}
+        filterComponent={
+          <TagFilterMenu
+            tags={availableTags}
+            selected={selectedTagIds}
+            setSelected={setSelectedTagIds}
+          />
+        }
       />
 
       <FlashList
@@ -410,13 +415,6 @@ export default function AllIngredientsScreen() {
             </Text>
           </View>
         }
-      />
-      <TagFilterModal
-        visible={tagModalVisible}
-        onClose={() => setTagModalVisible(false)}
-        tags={availableTags}
-        selected={selectedTagIds}
-        setSelected={setSelectedTagIds}
       />
     </View>
   );
